backend/pyproject.toml
[project]
name = "Zrata-X"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "celery>=5.6.0",
    "celery-redbeat>=2.3.3",
    "crawl4ai>=0.7.8",
    "fastapi>=0.127.1",
    "httpx>=0.28.1",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "orjson>=3.11.5",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "supermemory>=3.12.1",
    "uvicorn[standard]>=0.40.0",
    "yfinance>=1.0",
]
